refactor(launch): extract the three inline launch package-context literals
Three launch families built their PackageOpenInfo inline at the click site,
so nothing outside the component could call the same derivation:

- context-panel repeated one identical literal at all four resolved
  recommends/suggests nav-link handlers; they now share an exported
  packageInfoForNavLink.
- My Learning's App Platform path member and Discover More handlers now
  call packageInfoForPathMember / packageInfoForDiscoverItem from a new
  pure src/learning-paths/launch-package-info module, which also takes
  over Discover More's manifest parse from useDiscoverMore.

Behaviour-preserving throughout: every factory returns exactly what the
literal it replaced returned, including the fields it does not set. The
existing MyLearningTab packageInfo assertions still run against the real
factories (the barrel mocks requireActual them), so they remain evidence
that the output is unchanged.

DiscoverMoreItem moves to types/learning-paths.types.ts to keep
launch-package-info and useDiscoverMore acyclic, and is re-exported from
useDiscoverMore so no consumer import changes.
